Am facut refactor la "Amount" in toate modulele si eventurile. Acum peste tot se va folosi forma corecta a sa din LOng in BigDecimal si invers si calcularea corecta a rutei

+++ payment-routing/src/main/java/com/zaheudev/routing/entity/RoutingCost.java
@@ -27,8 +27,8 @@ public class RoutingCost {
      * @param amount
      * @return
      */
-    public BigDecimal calculateTotalCost(Long amount) {;
+    public BigDecimal calculateTotalCost(BigDecimal amount) {;
         //i think this is the formula, but it can be changed if needed
-        return (percentageFee.multiply(BigDecimal.valueOf(amount))).add(fixedFee).add(BigDecimal.valueOf(authorizationRate));
+        return (percentageFee.multiply(amount)).add(fixedFee).add(BigDecimal.valueOf(authorizationRate));
     }
 }
